Add support for SW5e
Hi, I am a member of the SW5e system development team. We were hoping to get this system added to your module! We're built off a base of D&D5e, so it should be more or less identical. If there is anything else we need to do, or if you have any questions, please feel free to reach out!

Thank you for creating an excellent module!